+++ app/src/main/java/org/sopt/pingle/util/view/ViewExt.kt
@@ -14,12 +14,3 @@ fun View.colorOf(@ColorRes resId: Int) = ContextCompat.getColor(context, resId)
 
 fun View.setBackgroundTint(@ColorRes resId: Int) =
     ViewCompat.setBackgroundTintList(this, ContextCompat.getColorStateList(context, resId))
-
-class ItemDiffCallback<T : Any>(
-    val onItemsTheSame: (T, T) -> Boolean,
-    val onContentsTheSame: (T, T) -> Boolean
-) : DiffUtil.ItemCallback<T>() {
-    override fun areItemsTheSame(oldItem: T, newItem: T): Boolean = onItemsTheSame(oldItem, newItem)
-    override fun areContentsTheSame(oldItem: T, newItem: T): Boolean =
-        onContentsTheSame(oldItem, newItem)
-}
